backend/requirements.txt
# Core dependencies
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
openai==1.12.0

# API and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Data processing
pandas==2.1.4
numpy==1.25.2

# Type checking and validation
pydantic==2.5.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: For better performance
orjson==3.9.10 